fix: Search hit return

+++ aisearch/src/main/java/com/elastic/aisearch/dto/SearchResponseDTO.java
@@ -3,7 +3,7 @@ package com.elastic.aisearch.dto;
 import java.util.List;
 
 public record SearchResponseDTO(
-    Integer hits,    
+    Long hits,
     Integer pages,
     Float timeTaken,
     List<SearchResultDTO> results
